Refactor App.js by splitting into modules
