+++ src/App.js
@@ -3,522 +3,16 @@ import React, { useState, useRef, useEffect, useLayoutEffect } from "react";
 import jsPDF from "jspdf";
 import html2canvas from "html2canvas";
 
-// --- STILDEFINITIONEN ---
-const styles = {
-  container: isMobile => ({
-    maxWidth: 600,
-    margin: "0 auto",
-    padding: isMobile ? "0 12px" : "0 24px",
-    overflowAnchor: "none"
-  }),
-  topBar: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    padding: "12px 0"
-  },
-  title: {
-    textAlign: "center",
-    margin: "8px 0 24px",
-    fontSize: 28,
-    fontWeight: 700
-  },
-  input: {
-    flex: 1,
-    padding: "10px 12px",
-    fontSize: 16,
-    borderRadius: 6,
-    border: "1px solid #ccc",
-    boxSizing: "border-box",
-  },
-  smallInput: {
-    padding: "8px 10px",
-    fontSize: 16,
-    WebkitTextSizeAdjust: "100%",
-    borderRadius: 6,
-    border: "1px solid #ccc",
-    minWidth: '70px',
-    boxSizing: "border-box",
-  },
-  textarea: {
-    width: "100%",
-    padding: "8px",
-    fontSize: 16,
-    WebkitTextSizeAdjust: "100%",
-    borderRadius: 6,
-    border: "1px solid #ccc",
-    marginTop: 8,
-    resize: "none",
-    overflow: "hidden",
-    overflowWrap: "break-word",
-    whiteSpace: "pre-wrap",
-    boxSizing: "border-box"
-  },
-  buttonPrimary: {
-    padding: "12px 0",
-    fontSize: 16,
-    borderRadius: 6,
-    border: 0,
-    background: "#388e3c",
-    color: "#fff",
-    cursor: "pointer",
-    width: "100%"
-  },
-  buttonSecondary: bg => ({
-    padding: "8px 16px",
-    fontSize: 14, // Standard-Fontsize für secondary buttons
-    borderRadius: 6,
-    border: 0,
-    background: bg,
-    color: "#fff",
-    cursor: "pointer"
-  }),
-  symptomAddButton: bg => ({
-    padding: "3px 10px",
-    fontSize: 16,
-    borderRadius: 6,
-    border: 0,
-    background: bg,
-    color: "#fff",
-    cursor: "pointer",
-    flexShrink: 0,
-    boxSizing: 'border-box'
-  }),
-  entryCard: (dark, isSymptomOnly = false) => ({
-    position: 'relative',
-    marginBottom: 16,
-    marginLeft: 5,
-    padding: 12,
-    borderRadius: 8,
-    background: isSymptomOnly
-      ? (dark ? "#3c3c46" : "#f0f0f5")
-      : (dark ? "#2a2a32" : "#fff"),
-    boxShadow: "0 1px 4px #0002",
-  }),
-  groupHeader: (isPdf) => ({
-    fontSize: 18,
-    fontWeight: 600,
-    margin: isPdf ? "32px 0 12px" : "24px 0 8px",
-    color: isPdf ? '#333' : undefined,
-  }),
-  dayCover: (dark) => ({
-    fontSize: 18,
-    fontWeight: 600,
-    margin: "24px 0 8px",
-    textAlign: 'center',
-    padding: '12px 0',
-    borderRadius: 8,
-    background: dark ? '#3a3a42' : '#e0e0e0',
-    cursor: 'pointer'
-  }),
-  toast: {
-    position: "fixed",
-    top: 16,
-    right: 16,
-    background: "#333",
-    color: "#fff",
-    padding: "8px 12px",
-    borderRadius: 4,
-    opacity: 0.9,
-    zIndex: 1000
-  },
-  backButton: {
-    padding: "6px 12px",
-    fontSize: 14,
-    borderRadius: 6,
-    border: 0,
-    background: "#1976d2",
-    color: "#fff",
-    cursor: "pointer"
-  },
-  glassyIconButton: (dark) => ({
-    background: dark ? 'rgba(255, 255, 255, 0.08)' : 'rgba(0, 0, 0, 0.04)',
-    border: dark ? '1px solid rgba(255, 255, 255, 0.15)' : '1px solid rgba(0, 0, 0, 0.1)',
-    borderRadius: 6,
-    padding: "6px 8px",
-    cursor: "pointer",
-    fontSize: 16,
-    lineHeight: 1,
-    color: dark ? '#f0f0f8' : '#333',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  }),
-  deleteIcon: {
-    position: 'absolute',
-    bottom: '8px',
-    right: '8px',
-    background: 'transparent',
-    border: 'none',
-    color: '#d32f2f',
-    fontSize: '20px',
-    cursor: 'pointer',
-  },
-  tagMarkerBase: {
-    position: 'absolute',
-    bottom: 0,
-    right: 0,
-    width: 0,
-    height: 0,
-    borderStyle: 'solid',
-    zIndex: 5,
-  },
-  tagMarkerOuter: (tagColor) => ({
-    ...styles.tagMarkerBase,
-    borderWidth: '0 0 28px 28px',
-    borderColor: `transparent transparent ${tagColor} transparent`,
-    cursor: 'pointer',
-  }),
-  tagMarkerInnerHint: (cardBgColor) => ({
-    ...styles.tagMarkerBase,
-    borderWidth: '0 0 16px 16px',
-    borderColor: `transparent transparent ${cardBgColor} transparent`,
-    pointerEvents: 'none',
-    zIndex: 6,
-  }),
-  colorPickerPopup: (dark) => ({
-    position: 'absolute',
-    bottom: '30px',
-    right: '5px',
-    background: dark ? '#4a4a52' : '#fff',
-    padding: '8px',
-    borderRadius: '6px',
-    boxShadow: '0 2px 10px rgba(0,0,0,0.25)',
-    zIndex: 30,
-    display: 'flex',
-    gap: '8px',
-  }),
-  colorPickerItem: (color, isActive, currentThemeDark) => ({
-    width: '24px',
-    height: '24px',
-    backgroundColor: color,
-    borderRadius: '4px',
-    cursor: 'pointer',
-    border: isActive ? (currentThemeDark ? '2px solid #FFFFFF' : '2px solid #000000') : '2px solid transparent',
-    boxSizing: 'border-box',
-  }),
-  pinContainer: {
-    position: 'absolute',
-    left: '0px',
-    top: 0,
-    bottom: 0,
-    width: '20px',
-    pointerEvents: 'none',
-  },
-  pin: (selected) => ({
-    position: 'absolute',
-    bottom: '2px',
-    left: '2px',
-    width: '16px',
-    height: '16px',
-    cursor: 'pointer',
-    pointerEvents: 'auto',
-    zIndex: 5,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    fontSize: '14px',
-  }),
-  connectionSvg: {
-    position: 'absolute',
-    left: '0px',
-    width: '20px',
-    pointerEvents: 'auto',
-    overflow: 'visible',
-    cursor: 'pointer',
-  },
-};
-
-// --- GLOBALE KONSTANTEN & FARBMAPPINGS ---
-
-const SYMPTOM_CHOICES = [
-  "Bauchschmerzen","Durchfall","Blähungen","Hautausschlag",
-  "Juckreiz","Schwellung am Gaumen","Schleim im Hals",
-  "Niesen","Kopfschmerzen","Rötung Haut"
-];
-const TIME_CHOICES = [
-  { label: "sofort", value: 0 }, { label: "nach 5 min", value: 5 },
-  { label: "nach 10 min", value: 10 }, { label: "nach 15 min", value: 15 },
-  { label: "nach 30 min", value: 30 }, { label: "nach 45 min", value: 45 },
-  { label: "nach 60 min", value: 60 }, { label: "nach 1,5 h", value: 90 },
-  { label: "nach 3 h", value: 180 }
-];
-
-const TAG_COLORS = {
-  GREEN: 'green',
-  RED: 'red',
-  YELLOW: 'yellow',
-  BROWN: '#c19a6b', // light brown for bowel movement
-};
-const TAG_COLOR_NAMES = {
-  [TAG_COLORS.GREEN]: "Standard",
-  [TAG_COLORS.RED]: "Symptome",
-  [TAG_COLORS.YELLOW]: "Vorgeschichte",
-  [TAG_COLORS.BROWN]: "Stuhlgang",
-};
-
-// --- HILFSFUNKTIONEN ---
-function resizeToJpeg(file, maxWidth = 800) {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onerror = (e) => { console.error("FileReader Error:", e); reject(reader.error); };
-    reader.onload = () => {
-      const img = new Image();
-      img.onerror = (e) => { console.error("Image Load Error:", e); reject(new Error("Bild konnte nicht als Bild interpretiert werden")); };
-      img.onload = () => {
-        try {
-          const scale = Math.min(1, maxWidth / img.width);
-          const w = img.width * scale;
-          const h = img.height * scale;
-          const c = document.createElement("canvas");
-          c.width = w;
-          c.height = h;
-          const ctx = c.getContext("2d");
-          if (!ctx) {
-            console.error("Canvas Context nicht erhalten");
-            return reject(new Error("Canvas Context Fehler"));
-          }
-          ctx.drawImage(img, 0, 0, w, h);
-          resolve(c.toDataURL("image/jpeg", 0.8));
-        } catch (canvasError) {
-          console.error("Canvas Error:", canvasError);
-          reject(new Error("Fehler bei der Bildverkleinerung (Canvas)"));
-        }
-      };
-      img.src = reader.result;
-    };
-    reader.readAsDataURL(file);
-  });
-}
-
-const getStrengthColor = (strengthVal) => {
-    const s = parseInt(strengthVal);
-    switch (s) {
-        case 1: return 'hsl(120, 65%, 50%)';
-        case 2: return 'hsl(35, 90%, 55%)';
-        case 3: return 'hsl(0, 75%, 55%)';
-        default:
-            if (s && s >= 3) return 'hsl(0, 75%, 55%)';
-            return 'hsl(120, 65%, 50%)';
-    }
-};
-
-const now = () => {
-  const d = new Date();
-  const day = String(d.getDate()).padStart(2, '0');
-  const month = String(d.getMonth() + 1).padStart(2, '0');
-  const year = d.getFullYear();
-  // Force 24h format to ensure parseDateString works regardless of locale
-  const time = d.toLocaleTimeString([], {
-    hour: "2-digit",
-    minute: "2-digit",
-    hour12: false,
-  });
-  return `${day}.${month}.${year} ${time}`;
-};
-const vibrate = (pattern) => {
-  try {
-    if (navigator.vibrate) navigator.vibrate(pattern);
-  } catch {}
-};
-
-
-const getTodayDateString = () => {
-  const d = new Date();
-  const day = String(d.getDate()).padStart(2, '0');
-  const month = String(d.getMonth() + 1).padStart(2, '0');
-  const year = d.getFullYear();
-  return `${day}.${month}.${year}`;
-};
-
-const parseDateString = (dateStr) => {
-    if (!dateStr || typeof dateStr !== 'string') return new Date(0);
-    const [datePart, timePart] = dateStr.split(' ');
-    if (!datePart || !timePart) return new Date(0);
-    const dateComponents = datePart.split('.').map(Number);
-    const timeComponents = timePart.split(':').map(Number);
-    if (dateComponents.length !== 3 || timeComponents.length !== 2) return new Date(0);
-    if ([...dateComponents, ...timeComponents].some(isNaN)) return new Date(0);
-    const [day, month, year] = dateComponents;
-    const [hour, minute] = timeComponents;
-    if (year < 1000 || year > 3000 || month < 1 || month > 12 || day < 1 || day > 31) return new Date(0);
-    if (hour < 0 || hour > 23 || minute < 0 || minute > 59) return new Date(0);
-    return new Date(year, month - 1, day, hour, minute);
-};
-
-const toDateTimePickerFormat = (displayDateStr) => {
-    if (!displayDateStr || typeof displayDateStr !== 'string') return "";
-    const [datePart, timePart] = displayDateStr.split(' ');
-    if (!datePart || !timePart) return "";
-    const dateComponents = datePart.split('.');
-    if (dateComponents.length !== 3) return "";
-    const [day, month, year] = dateComponents.map(s => String(s).padStart(2,'0'));
-    const timeParts = timePart.split(':').map(s => String(s).padStart(2,'0'));
-    if (timeParts.length !== 2) return "";
-    if (isNaN(parseInt(year)) || isNaN(parseInt(month)) || isNaN(parseInt(day)) || isNaN(parseInt(timeParts[0])) || isNaN(parseInt(timeParts[1]))) return "";
-    if (parseInt(month) < 1 || parseInt(month) > 12 || parseInt(day) < 1 || parseInt(day) > 31 || parseInt(timeParts[0]) < 0 || parseInt(timeParts[0]) > 23 || parseInt(timeParts[1]) < 0 || parseInt(timeParts[1]) > 59) return "";
-    return `${year}-${month}-${day}T${timeParts[0]}:${timeParts[1]}`;
-};
-
-const fromDateTimePickerFormat = (pickerDateStr) => {
-    if (!pickerDateStr || typeof pickerDateStr !== 'string') return "";
-    const [datePart, timePart] = pickerDateStr.split('T');
-    if (!datePart || !timePart) return "";
-    const dateComponents = datePart.split('-');
-    if (dateComponents.length !== 3) return "";
-    const [year, month, day] = dateComponents;
-    const timeParts = timePart.split(':');
-    if (timeParts.length !== 2) return "";
-    if (isNaN(parseInt(year)) || isNaN(parseInt(month)) || isNaN(parseInt(day)) || isNaN(parseInt(timeParts[0])) || isNaN(parseInt(timeParts[1]))) return "";
-    if (parseInt(month) < 1 || parseInt(month) > 12 || parseInt(day) < 1 || parseInt(day) > 31 || parseInt(timeParts[0]) < 0 || parseInt(timeParts[0]) > 23 || parseInt(timeParts[1]) < 0 || parseInt(timeParts[1]) > 59) return "";
-    return `${String(day).padStart(2, '0')}.${String(month).padStart(2, '0')}.${year} ${String(timeParts[0]).padStart(2, '0')}:${String(timeParts[1]).padStart(2, '0')}`;
-};
-
-const sortSymptomsByTime = (symptoms) => {
-    return [...symptoms].sort((a, b) => {
-        const tA = typeof a.time === 'number' ? a.time : parseFloat(a.time) || 0;
-        const tB = typeof b.time === 'number' ? b.time : parseFloat(b.time) || 0;
-        if (tA === tB) return (a.txt || '').localeCompare(b.txt || '');
-        return tA - tB;
-    });
-};
-
-// --- UI UTILITY KOMPONENTEN ---
-const PdfButton = ({ onClick }) => (
-  <button onClick={onClick} className="haptic" title="Export PDF" style={styles.buttonSecondary("#d32f2f")}>
-    PDF
-  </button>
-);
-const InsightsButton = ({ onClick }) => (
-  <button onClick={onClick} className="haptic" title="Insights" style={styles.buttonSecondary("#1976d2")}>
-    Insights
-  </button>
-);
-const BackButton = ({ onClick }) => (
-  <button onClick={onClick} className="haptic" title="Zurück" style={styles.backButton}>← Zurück</button>
-);
-const CameraButton = ({ onClick }) => (
-  <button onClick={onClick} className="haptic" title="Foto" style={{
-    width: 36, height: 36, borderRadius: 6, border: 0,
-    background: "#247be5", display: "flex", alignItems: "center",
-    justifyContent: "center", cursor: "pointer"
-  }}>📷</button>
-);
-const ImgStack = ({ imgs, onDelete }) => (
-  <div className="img-stack-container" style={{ display: "flex", alignItems: "center", marginBottom: 8 }}>
-    {imgs.map((src, i) => (
-      <div key={i} className="img-stack-item" style={{ position: "relative", marginLeft: i ? -12 : 0, zIndex: imgs.length - i }}>
-        <img
-          src={src}
-          alt={`entry_image_${i}`}
-          className="fade-img"
-          style={{
-            width: 40, height: 40, objectFit: "cover",
-            borderRadius: 6, border: "2px solid #fff",
-            boxShadow: "0 1px 4px #0003",
-            opacity: 0,
-            transition: 'opacity 0.3s ease'
-          }}
-          onLoad={e => { e.currentTarget.style.opacity = 1; }}
-          onError={e => { e.currentTarget.style.display = "none"; }}
-        />
-        {onDelete && (
-          <span onClick={() => onDelete(i)} style={{
-            position: "absolute", top: -6, right: -6,
-            background: "#c00", color: "#fff",
-            borderRadius: "50%", width: 18, height: 18,
-            display: "flex", alignItems: "center",
-            justifyContent: "center", fontSize: 12,
-            cursor: "pointer"
-          }}>×</span>
-        )}
-      </div>
-    ))}
-  </div>
-);
-
-const SymTag = ({ txt, time, strength, dark, onDel, onClick }) => {
-  const tagBackgroundColor = "#fafafa"; // Einheitliche Farbe für alle Symptome
-  const tagTextColor = "#1a1f3d";
-  const displayStrength = Math.min(parseInt(strength) || 1, 3);
-
-  return (
-    <div onClick={onClick} style={{
-      display: "inline-flex", alignItems: "center",
-      background: tagBackgroundColor,
-      color: tagTextColor,
-      borderRadius: 6, padding: "6px 10px",
-      margin: "3px 4px 3px 0", fontSize: 14,
-      cursor: onClick ? "pointer" : "default",
-      overflowWrap: "break-word", whiteSpace: "normal"
-    }}>
-      {strength && (
-        <span style={{
-            display: 'inline-flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            width: '16px',
-            height: '16px',
-            borderRadius: '50%',
-            backgroundColor: 'rgba(255, 255, 255, 0.2)',
-            color: '#333333',
-            fontSize: '10px',
-            fontWeight: 'bold',
-            marginRight: '5px',
-            lineHeight: 1,
-            flexShrink: 0,
-            border: `2px solid ${getStrengthColor(displayStrength)}`,
-            boxSizing: 'border-box',
-        }}>
-            {displayStrength}
-        </span>
-      )}
-      {txt}
-      <span style={{ marginLeft: 8, fontSize: 12, opacity: 0.8, flexShrink: 0 }}>
-        {TIME_CHOICES.find(t => t.value === time)?.label || `${time} min`}
-      </span>
-      {onDel && (
-        <span onClick={e => { e.stopPropagation(); onDel(); }} style={{
-          marginLeft: 8, cursor: "pointer",
-          fontSize: 16,
-          color: "#c00",
-          fontWeight: 700
-        }}>×</span>
-      )}
-    </div>
-  );
-};
-
-// --- DATENVERARBEITUNGSKOMPONENTEN (z.B. Insights) ---
-function Insights({ entries }) {
-  const map = {};
-  entries.forEach(e => {
-    (e.symptoms || []).forEach(s => {
-      if (!map[s.txt]) map[s.txt] = { count: 0, foods: {} };
-      map[s.txt].count++;
-      const foodKey = e.food || "(Kein Essen)";
-      map[s.txt].foods[foodKey] = (map[s.txt].foods[foodKey] || 0) + 1;
-    });
-  });
-  const sorted = Object.entries(map).sort((a, b) => b[1].count - a[1].count);
-  return (
-    <div>
-      <h2 style={{ textAlign: "center", margin: "16px 0" }}>Insights</h2>
-      {sorted.length === 0 && <p>Keine Symptome erfasst.</p>}
-      {sorted.map(([symptom, data]) => (
-        <div key={symptom} style={{ marginBottom: 24 }}>
-          <h3>{symptom} ({data.count})</h3>
-          <ul>
-            {Object.entries(data.foods).sort((a, b) => b[1] - a[1]).map(([food, cnt]) => (
-              <li key={food}>{food}: {cnt}</li>
-            ))}
-          </ul>
-        </div>
-      ))}
-    </div>
-  );
-}
-
+import styles from "./styles";
+import { SYMPTOM_CHOICES, TIME_CHOICES, TAG_COLORS, TAG_COLOR_NAMES } from "./constants";
+import { resizeToJpeg, getStrengthColor, now, vibrate, getTodayDateString, parseDateString, toDateTimePickerFormat, fromDateTimePickerFormat, sortSymptomsByTime } from "./utils";
+import PdfButton from "./components/PdfButton";
+import InsightsButton from "./components/InsightsButton";
+import BackButton from "./components/BackButton";
+import CameraButton from "./components/CameraButton";
+import ImgStack from "./components/ImgStack";
+import SymTag from "./components/SymTag";
+import Insights from "./components/Insights";
 // --- HAUPTANWENDUNGSKOMPONENTE: App ---
 export default function App() {
   // --- STATE VARIABLEN ---
