+++ src/components/Insights.js
@@ -1,0 +1,32 @@
+import React from "react";
+
+function Insights({ entries }) {
+  const map = {};
+  entries.forEach(e => {
+    (e.symptoms || []).forEach(s => {
+      if (!map[s.txt]) map[s.txt] = { count: 0, foods: {} };
+      map[s.txt].count++;
+      const foodKey = e.food || "(Kein Essen)";
+      map[s.txt].foods[foodKey] = (map[s.txt].foods[foodKey] || 0) + 1;
+    });
+  });
+  const sorted = Object.entries(map).sort((a, b) => b[1].count - a[1].count);
+  return (
+    <div>
+      <h2 style={{ textAlign: "center", margin: "16px 0" }}>Insights</h2>
+      {sorted.length === 0 && <p>Keine Symptome erfasst.</p>}
+      {sorted.map(([symptom, data]) => (
+        <div key={symptom} style={{ marginBottom: 24 }}>
+          <h3>{symptom} ({data.count})</h3>
+          <ul>
+            {Object.entries(data.foods).sort((a, b) => b[1] - a[1]).map(([food, cnt]) => (
+              <li key={food}>{food}: {cnt}</li>
+            ))}
+          </ul>
+        </div>
+      ))}
+    </div>
+  );
+}
+
+export default Insights;
