+++ src/styles.js
@@ -1,0 +1,226 @@
+const styles = {
+  container: isMobile => ({
+    maxWidth: 600,
+    margin: "0 auto",
+    padding: isMobile ? "0 12px" : "0 24px",
+    overflowAnchor: "none"
+  }),
+  topBar: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "12px 0"
+  },
+  title: {
+    textAlign: "center",
+    margin: "8px 0 24px",
+    fontSize: 28,
+    fontWeight: 700
+  },
+  input: {
+    flex: 1,
+    padding: "10px 12px",
+    fontSize: 16,
+    borderRadius: 6,
+    border: "1px solid #ccc",
+    boxSizing: "border-box",
+  },
+  smallInput: {
+    padding: "8px 10px",
+    fontSize: 16,
+    WebkitTextSizeAdjust: "100%",
+    borderRadius: 6,
+    border: "1px solid #ccc",
+    minWidth: '70px',
+    boxSizing: "border-box",
+  },
+  textarea: {
+    width: "100%",
+    padding: "8px",
+    fontSize: 16,
+    WebkitTextSizeAdjust: "100%",
+    borderRadius: 6,
+    border: "1px solid #ccc",
+    marginTop: 8,
+    resize: "none",
+    overflow: "hidden",
+    overflowWrap: "break-word",
+    whiteSpace: "pre-wrap",
+    boxSizing: "border-box"
+  },
+  buttonPrimary: {
+    padding: "12px 0",
+    fontSize: 16,
+    borderRadius: 6,
+    border: 0,
+    background: "#388e3c",
+    color: "#fff",
+    cursor: "pointer",
+    width: "100%"
+  },
+  buttonSecondary: bg => ({
+    padding: "8px 16px",
+    fontSize: 14, // Standard-Fontsize für secondary buttons
+    borderRadius: 6,
+    border: 0,
+    background: bg,
+    color: "#fff",
+    cursor: "pointer"
+  }),
+  symptomAddButton: bg => ({
+    padding: "3px 10px",
+    fontSize: 16,
+    borderRadius: 6,
+    border: 0,
+    background: bg,
+    color: "#fff",
+    cursor: "pointer",
+    flexShrink: 0,
+    boxSizing: 'border-box'
+  }),
+  entryCard: (dark, isSymptomOnly = false) => ({
+    position: 'relative',
+    marginBottom: 16,
+    marginLeft: 5,
+    padding: 12,
+    borderRadius: 8,
+    background: isSymptomOnly
+      ? (dark ? "#3c3c46" : "#f0f0f5")
+      : (dark ? "#2a2a32" : "#fff"),
+    boxShadow: "0 1px 4px #0002",
+  }),
+  groupHeader: (isPdf) => ({
+    fontSize: 18,
+    fontWeight: 600,
+    margin: isPdf ? "32px 0 12px" : "24px 0 8px",
+    color: isPdf ? '#333' : undefined,
+  }),
+  dayCover: (dark) => ({
+    fontSize: 18,
+    fontWeight: 600,
+    margin: "24px 0 8px",
+    textAlign: 'center',
+    padding: '12px 0',
+    borderRadius: 8,
+    background: dark ? '#3a3a42' : '#e0e0e0',
+    cursor: 'pointer'
+  }),
+  toast: {
+    position: "fixed",
+    top: 16,
+    right: 16,
+    background: "#333",
+    color: "#fff",
+    padding: "8px 12px",
+    borderRadius: 4,
+    opacity: 0.9,
+    zIndex: 1000
+  },
+  backButton: {
+    padding: "6px 12px",
+    fontSize: 14,
+    borderRadius: 6,
+    border: 0,
+    background: "#1976d2",
+    color: "#fff",
+    cursor: "pointer"
+  },
+  glassyIconButton: (dark) => ({
+    background: dark ? 'rgba(255, 255, 255, 0.08)' : 'rgba(0, 0, 0, 0.04)',
+    border: dark ? '1px solid rgba(255, 255, 255, 0.15)' : '1px solid rgba(0, 0, 0, 0.1)',
+    borderRadius: 6,
+    padding: "6px 8px",
+    cursor: "pointer",
+    fontSize: 16,
+    lineHeight: 1,
+    color: dark ? '#f0f0f8' : '#333',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  }),
+  deleteIcon: {
+    position: 'absolute',
+    bottom: '8px',
+    right: '8px',
+    background: 'transparent',
+    border: 'none',
+    color: '#d32f2f',
+    fontSize: '20px',
+    cursor: 'pointer',
+  },
+  tagMarkerBase: {
+    position: 'absolute',
+    bottom: 0,
+    right: 0,
+    width: 0,
+    height: 0,
+    borderStyle: 'solid',
+    zIndex: 5,
+  },
+  tagMarkerOuter: (tagColor) => ({
+    ...styles.tagMarkerBase,
+    borderWidth: '0 0 28px 28px',
+    borderColor: `transparent transparent ${tagColor} transparent`,
+    cursor: 'pointer',
+  }),
+  tagMarkerInnerHint: (cardBgColor) => ({
+    ...styles.tagMarkerBase,
+    borderWidth: '0 0 16px 16px',
+    borderColor: `transparent transparent ${cardBgColor} transparent`,
+    pointerEvents: 'none',
+    zIndex: 6,
+  }),
+  colorPickerPopup: (dark) => ({
+    position: 'absolute',
+    bottom: '30px',
+    right: '5px',
+    background: dark ? '#4a4a52' : '#fff',
+    padding: '8px',
+    borderRadius: '6px',
+    boxShadow: '0 2px 10px rgba(0,0,0,0.25)',
+    zIndex: 30,
+    display: 'flex',
+    gap: '8px',
+  }),
+  colorPickerItem: (color, isActive, currentThemeDark) => ({
+    width: '24px',
+    height: '24px',
+    backgroundColor: color,
+    borderRadius: '4px',
+    cursor: 'pointer',
+    border: isActive ? (currentThemeDark ? '2px solid #FFFFFF' : '2px solid #000000') : '2px solid transparent',
+    boxSizing: 'border-box',
+  }),
+  pinContainer: {
+    position: 'absolute',
+    left: '0px',
+    top: 0,
+    bottom: 0,
+    width: '20px',
+    pointerEvents: 'none',
+  },
+  pin: (selected) => ({
+    position: 'absolute',
+    bottom: '2px',
+    left: '2px',
+    width: '16px',
+    height: '16px',
+    cursor: 'pointer',
+    pointerEvents: 'auto',
+    zIndex: 5,
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    fontSize: '14px',
+  }),
+  connectionSvg: {
+    position: 'absolute',
+    left: '0px',
+    width: '20px',
+    pointerEvents: 'auto',
+    overflow: 'visible',
+    cursor: 'pointer',
+  },
+};
+
+export default styles;
